requirements.txt
flask
fastai
torch
torchvision
pillow
numpy<2.0
gunicorn